Turn off auto-save. Silence compiler warning by casting mask layer value in exporter.
git-svn-id: https://gforge.sci.utah.edu/svn/seg3d2/branches/aribex@1923 305cdda6-5ce1-45b3-a98d-dfc68c8b3305

+++ src/Application/LayerIO/NrrdLayerExporter.cc
@@ -160,7 +160,7 @@ bool NrrdLayerExporter::export_mask_label( const std::string& file_path )
     // TODO: temporary hack
     for (size_t k = 0; k < this->label_values_.size(); ++k)
     {
-      this->label_values_[k] = k;
+      this->label_values_[k] = static_cast<double>(k);
     }
   }
